Restructure and rename thread local things in std.
